refactor(agent): localize VM recovery attempt state
