+++ packages/agent/vitest.unit.config.ts
@@ -124,7 +124,9 @@ export default defineConfig({
       "test/chain-reconciler.test.ts",
       "test/vm-reconcile-self-prime.test.ts",
       "test/core-fills-gap.test.ts",
-      "test/vm-recovery-microbatch-adversarial.test.ts",
+      "test/vm-recovery-microbatch-planner.test.ts",
+      "test/vm-recovery-provider-policy.test.ts",
+      "test/vm-recovery-microbatch-host.test.ts",
       "test/vm-recovery-footprint-bridge.test.ts",
       "test/vm-recovery-footprint-bridge-adversarial.test.ts",
       "test/finalization-reconcile-negative-memo.test.ts",
